Create Vector2 class and Point2D extends it
